Drop a redundant cast, and gate on exit status not on grep
The lint is trivial. The reason it reached a push is not: I was gating
the commit on a pipeline ending in grep, and grep that finds errors
exits successfully, so && carried on past a failing clippy. Checking
cargo own exit status is the only version of that gate which works.

+++ crates/runity/src/render.rs
@@ -1366,7 +1366,7 @@ impl Renderer {
                     pass.set_bind_group(
                         2,
                         &self.pose_bind_group,
-                        &[*pose as u32 * self.pose_stride as u32],
+                        &[*pose * self.pose_stride as u32],
                     );
                     pass.set_vertex_buffer(0, mesh.vertices.slice(..));
                     pass.set_vertex_buffer(1, self.instances.slice(..));
